+++ Assets/Scripts/Player/Standard_Projectile.cs
@@ -8,7 +8,7 @@ using UnityEngine;
 public class Standard_Projectile : MonoBehaviour {
     /// <summary>This projectiles damage</summary>
     [SerializeField]
-    private int damage = 2;
+    private byte damage = 2;
 
     ///<summary>References the bullets attached Teamhandler</summary>
     [SerializeField]
